Fix up legacy code to use new interfaces This took way more time than it was worth. Converted line endings to Unix, some other trivial crap. Thanks to Phoenix616 for already having a multi-version pom, it would have taken me much longer to figure out without a functional example.
The only legacy code which is not supported is that of the pre-versioned package era. I compiled a copy of Craftbukkit about 15 commits prior to the introduction of versioned packages, and NMS had already been fiddled with. Anyone who needs support for builds in that range can use an older OpenInv build instead, it's not like UUIDs are a big concern.